Peggle Deluxe: Time start fix, comment clarity
Previously, the timer would erroneously start when Peggle was launched
because the memory hadn't yet been properly initialised. Now we check to
make sure that the StoryScreen object is initialised correctly by
testing that it links back to the parent object in the right spot.

Comments also needed a little bit of clarification.

(The previous commit, 46e9d5cc, was successfully used on a complete,
verified run of Adventure Mode in 2:41:20 between the time of that
commit and this one. Video and verification at
<http://www.speedrun.com/run/nz1pde7z>.)